Add more classes
Introduced an artificial delay when unlocking alongside some new classes that allow for fancy css animations to happen.
This is absolutely awful code wise but you aren't looking at this.. right?
No worries i'll clean it up one day.

At the moment there is no way to disable the delay nor is there an example of how to add animations, To be added later.